requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
tensorflow-cpu==2.19.0
numpy==1.26.4
pillow==10.4.0
werkzeug==3.0.3
jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
blinker==1.6.2
sqlalchemy==2.0.37
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
google-cloud-storage==2.14.0 